Fix `BigInt#testBit` to throw `ArithmeticException` when `n` is negative.
The `_bigInteger` field, a mutable value initialized by `BigInt#bigInteger`, is accessed directly by `BigInt#testBit` when the input `n` is negative.
As a result, `testBit` currently throws a `NullPointerException`.

The expected behavior is for `testBit` to throw an `ArithmeticException` when `n` is negative, as specified by the underlying `java.math.BigInteger` API:
https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/math/BigInteger.html#testBit(int)

This issue was not detected in `BigIntTest` because `_bigInteger` is initialized by earlier tests.
Additionally, using `BigInt(42).testBit(-3)` for testing does not reproduce the issue, as `BigInt(42)` uses a cached instance where `_bigInteger` is already initialized.

See: https://github.com/scala-native/scala-native/issues/2897