`substSym` replaces owners of `RefinementClassSymbol`s
Given

```
class C {
  def f = new {
    def g = new {
      def h = 1
    }
  }
}
```

the type of `C.f` is `AnyRef { def g: AnyRef { def h: Int } }`.

The inner `RefinedType` has as `typeSymbol` a `RefinementClassSymbol`. Its
owner points to `<$anon>.g`, the symbol for method `g` in the anonymous class.

Thus, the anonymous class symbol escapes its scope, the type of `f` should
not contain any references to it.

Instead, the `RefinementClassSymbol` of the inner refinement should have
symbol `g` from the scope of the outer refinement as its owner.

---------

Context: there's reasons to think that this can cause overcompilation in zinc.
Or maybe undercompilation.

In the xsbt-dependency phase, when a reference to `f` is processed, the phase
calls `addTypeDependencies` on the the type of `f`. The `addClassDependency`
method goes to the enclosing class, that way we end up adding a dependency
on the `<$anon>` class.

We get to `externalDependency` in zinc and eventually to `lookupAnalyzedClass`
https://github.com/sbt/zinc/blob/v1.10.4/internal/zinc-core/src/main/scala/sbt/internal/inc/Lookup.scala#L56

Here `analysis.relations.productClassName.reverse(binaryClassName)` finds
nothing as `relations.productClassName` doesn't contain anonymous classes.